修改为int

+++ response/request_paging.go
@@ -4,17 +4,17 @@ const maxPageSize = 100
 const minPageSize = 20
 
 type RequestPaging struct {
-	PageSize int64 `form:"pageSize"` // 大小
-	Current  int64 `form:"current"`  // 页码
+	PageSize int `form:"pageSize"` // 大小
+	Current  int `form:"current"`  // 页码
 }
 
 // 偏移
-func (req *RequestPaging) Offset() int64 {
+func (req *RequestPaging) Offset() int {
 	return (req.GetCurrent() - 1) * req.GetPageSize()
 }
 
 // 分页大小
-func (req *RequestPaging) GetPageSize() int64 {
+func (req *RequestPaging) GetPageSize() int {
 	if req.PageSize > maxPageSize {
 		return maxPageSize
 	}
@@ -26,7 +26,7 @@ func (req *RequestPaging) GetPageSize() int64 {
 }
 
 // 当前页码
-func (req *RequestPaging) GetCurrent() int64 {
+func (req *RequestPaging) GetCurrent() int {
 	if req.Current == 0 {
 		return 1
 	}
